changed the cost property from an IntegerProperty to FloatProperty in class DeclarationLine the same applies on the max_cost property of the class DeclarationSubType

+++ Ilmoitus/ilmoitus_model.py
@@ -213,7 +213,7 @@ class DeclarationType(ndb.Model):
 # DeclarationLine Model class
 class DeclarationLine(ndb.Model):
     receipt_date = ndb.StringProperty()  # DateProperty?
-    cost = ndb.IntegerProperty()
+    cost = ndb.FloatProperty()
     declaration_sub_type = ndb.KeyProperty(kind=DeclarationSubType)
 
     def get_object_as_data_dict(self):
